Fix for file tree braking
The safety in doCreate may be unnecessary. What this will do is the
cash/index will be checked beforehand thus refreshing the cache. What
happens I think is that Brackets does not do anything with the file
index when it recovers from a previous filesystem error; really I don’t
think it messes with the Brackets filesystem in general so it does not
reflect the actual filesystem on the computer. This creates a lag from
previous errors in that the filesystem on Brackets would contain
nonexistent files and folders.